move networking plugins out of main index
